requirements.txt
torch
openai-whisper
transformers
pydub
moviepy
nltk
spacy
tqdm
numpy
ffmpeg-python
